Mark Razor source generator files as generated
Give the output files from the Razor Source Generator a ".g" suffix so that
they are treated as a generated file to prevent it being flagged by source
analyzers for violations the application developer cannot fix themselves and
so that tools such as coverlet ignore them.
Relates to dotnet/aspnetcore#31719.