Do not catch exceptions thrown as part of applying updates
In the linked issue, applying a hot reload update would throw an error. However the error is caught and logged which is not displayed to the user by default.
In addition since the exception is caught, dotnet-watch assumes the apply succeeded and reports as such. Future hot reload updates will fail since they are working
off on an incorrect baseline. Removing the catch works since watch will not receive a success signal and treat it as a failure.

Contributes to https://github.com/dotnet/aspnetcore/issues/34446