Vary the cache key for TagHelpersFromReferences on SuppressRazorSourceGenerator (#20920)
RSG is invoked with 2 sets of values for SuppressRazorSourceGenerator first with it being set, and once without.
When set, RSG, does not compute any tag helpers from references and produces an empty result. Prior to this change,
we were caching this value invariant of SuppressRazorSourceGenerator. Consequently the next time the RSG was invoked,
the source generator infrastructure would assume that the empty result is still valid and use it for codegen.

The result of the bug is that in a hot reload compilation, none of the tag helpers were resolved and razor would treat tag helpers
from references as markup text. We fix this scenario by varying TagHelpersFromReferences using SuppressRazorSourceGenerator. Additionally
added a test to validate that we're generating the expected source.