Add a static cache for GetAssemblyVersion
In real-world projects like OrchardCore, ResolvePackageFileConflicts can
run on many files. Since it doesn't cache at all today, that can result
in reading assembly metadata out of the files many, many times.

Instead, cache the results of the call within GetAssemblyVersion and
check only the timestamp of the file on subsequent reads. Even that is
probably too conservative--we could consider caching files in the .NET
SDK directory "forever" (the lifetime of the MSBuild node), and possibly
expanding that to the NuGet packages directory as well, since its
contents should be immutable (but users do occasionally change them).

For this point in the release cycle I chose to be conservative and do
the timestamp check.