Include dotnethome WIXOBJ files during light command package drop creation (#12100)
After https://github.com/dotnet/installer/pull/11843, post build signing started failing, failing to find one of the input wixobj files. The wixobj was not included in the wixpack. This is because the installer repo uses custom scripts to generate MSIs, rather than the arcade tasks (which would handle adding this transparently), so the input files need to be added explicitly.