Prevent GlobalUsings file from being cleaned.
Performing a clean removes all files that were recorded using FileWrites.
In VS performing a clean and doing nothing starts producing build warnings for missing namespaces since the GlobalUsings file
has gone amiss. This change ensures that the file is not recorded and consequently not removed as part of "Clean".