Fix ILLink test instalibility
The tests created a NuGet package and then refered to it from other test projects with package reference. But some tests recreated the nuget package with slightly different content. Nuget doesn't always "re-restore" the nuget package it seems, and in the local nuget cache some tests got stale version of the package, causing the tests to fail.

Depending on the order in which the tests ran this could lead to failure.

Also reenabled a test which was disabled due to an already fixed bug.