Assume basline capabilities
In a Blazor app, the apply-update capabilities are available after the app is up and running in the browser.
Occasionally it takes long for the app to start up and our baseline task times out. Currently we return an empty
list of capabilities which causes the compiler to produce no deltas. This PR bumps up the timeout ever so slightly
and returns baseline capabilities instead.

Fixes https://github.com/dotnet/aspnetcore/issues/36723