Filter out project early if it doesn't have any formattable documents
Improves the runtime in large solutions when a small number of files is
included (`--include`). This is a common use-case for developers to
lint/format only modified files in their branch.

In a solution with about 80 projects, this improves the runtime by about
50% when `--include SomeProject/SomeFile.cs` is specified.

Related issues: https://github.com/dotnet/format/issues/1378 and
https://github.com/dotnet/format/issues/757