Allow skipping ILLink target
For NativeAOT we need to be able to opt in to all `PublishTrimmed` behaviors except the part that runs ILLink. We currently do that in a hacky way (set `PublishTrimmed` to true and create ILLink semaphore file to trick ILLink into not running); it's not great.

I tried introducing a `EnableTrimmingBehaviors` property that becomes true when `PublishTrimmed` is true, but in the end, we would like to be able to support NativeAOT for projects that have `PublishTrimmed` in the project file without NativeAOT failing, so what we really need is a way to disable running ILLink even though `PublishTrimmed` is true.