Update expected warnings in ILLink test
The new version of ILLink slightly changes the formatting of warning messages (we add space after comma in a method signature). This has very low probablity of actually breaking anybody, but the test infra compares strings directly, so it noticed.
This change updates the expected warnings with the new format.

It also removes an expected warning which is no longer generated.