Support multiple runtime packs per FrameworkReference
The `<ProcessFrameworkReferences/>` task is already able to process a
`@(KnownFrameworkReference)` with multiple `%(RuntimePackNamePatterns)`
defined in a semi-colon delimited list.  However, if multiple runtime
pack patterns are used the `<ResolveFrameworkReferences/>` will fail
soon after with the following error:

    C:\Program Files\dotnet\sdk\6.0.300-preview.22173.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(326,5): error MSB4018: The "ResolveFrameworkReferences" task failed unexpectedly.
    C:\Program Files\dotnet\sdk\6.0.300-preview.22173.2\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(326,5): error MSB4018: System.ArgumentException: An item with the same key has already been added. Key: Microsoft.Android

Fixes `<ResolveFrameworkReferences/>` to allow multiple runtime packs to
be associated with a single `@(FrameworkReference)`.  The following
`@(ResolvedFrameworkReferences)` metadata will now contain semi-colon
delimited lists of runtime pack information if multiple patterns are
provided:

    %(ResolvedFrameworkReferences.RuntimePackPath) = path1;path2
    %(ResolvedFrameworkReferences.RuntimePackName) = pack1;pack2.rid
    %(ResolvedFrameworkReferences.RuntimePackVersion) = 7.0.0;7.0.0