Update Microsoft.NET.Compatibility.Common.targets
Fixes https://github.com/dotnet/sdk/issues/24943

Make sure that the `BuildProjectReferences` property  is set to false when the `NoBuild=true` property is set, to avoid the ResolveReferences target to build project references when it shouldn't.

Before this change, a non cross-targeting build avoided to use an MSBuild task to gather the project's references but as NuGet's Pack target itself already doesn't differentiate between if it's running as part of a cross targeting or a non-cross targeting build and always calls into an MSBuild task, this optimization isn't necessary in package validation. Instead, this change brings package validation in sync with how NuGet's Pack target invokes inner build projects to make sure that additional project evaluations are avoided.