Move logger into ApiCompatibility and extend them
Move the loggers into ApiCompatibility as logging isn't
PackageValidation specific. This allows to expose the ApiCompatibility
as an task be able to use logging and generate the suppression
file.

Update the resource files becaues of the move.

Change the ICompatibilityLogger type from an interface to an abstract
class so that it can serve a default implementation for generating the
suppression file. Also change the return type of the
LogError/LogWarning methods to boolean so that callees know if the
message was actually logged.