Move common frontend code into Shared, Validators
Move common frontend code (initialization and execution of the
validators) into a shared project. Update solution files.

Add some properties to make the console project packable and consumeable
as a global tool.

Update the design of the validators and define an interface so that they
can work with multiple packages passed in via multiple Validate method
invocations.