Support suppressions without a diagnostic id
Any type that wants to create/log a suppression, needs to continue to
provide a suppression id, in order for suppressions to be uniquely
identifyable and suppressable (ie via NoWarn). Therefore the suppression
id is non nullable in the Suppression type.

A customer can provide a suppression file that contains a suppression
without a diagnostic id. When that is being deserialized, a Suppression
object with an empty diagnostic id is created. This change updates the
suppression engine to support this new form of global suppression, for
which only a left and right (+ IsBaselineSuppression) is passed in.

Adding tests that verify that this added global suppression type is
respected and correctly roundtrips.