.NET Framework tasks reference older MSBuild
This fixes a class of break in the scenario "New .NET SDK running on old
Visual Studio/ MSBuild.exe" by ensuring that the SDK task assemblies
don't get references to newer versions of MSBuild's transitive
references than are available in the current MSBuild environment, even
if that's older.

We have to specify explicit old versions of some `System.*` assemblies
to get versions that will be consistent with the older versions of
MSBuild.