Pass in reference assemblies to ApiCompat
This fixes APICompat's target which incorrectly was feeding in the ReferencePath item which contains implementation assemblies. Instead use the ReferencePathWithRefAssemblies item which contains the actual reference assemblies. This is also the input that the compiler receives.

As an example, System.Threading.ThreadPool's implementation assembly was building against ref/System.Private.CoreLib.dll but APICompat received the src/System.Private.CoreLib.dll assembly.