Fix single assembly mode when invoking ApiComparer
Use the correct overload when invoking ApiComparer to compare left and
right assemblies, and assembly sets with each other. Also fix the
underlying ApiCompatRunnerWorkItem type to allow multiple right assembly
sets for a given left assembly set.

Update a few tests because of the data structure change in the
ApiCompatRunnerWorkItem type.