Clean up Razor Server pid file if the path doesn't exist
When the server path in the pid file didn't exist we'd hit an error:

```
The shutdown command failed: The application to execute does not exist
```

This can happen when the file wasn't cleaned up and the SDK was uninstalled in the meantime.
Ensure we delete the pid file in that case and don't error.

Fixes https://github.com/dotnet/sdk/issues/10573