Make TryGetInternalDefaultConstructor work without internal metadata
TryGetInternalDefaultConstructor was previously expecting to be able to
see an internal constructor in order to generate a non-public default
constructor.  This will not be the case if the INamedTypeSymbol came
from an assembly imported with MetadataImportOptions.Public (as is the
default) or if the linker ran on the assembly and removed the constructor.

This change adds the default constructor when none are present and fixes
the method to do nothing for types which do not get default constructors.