[APICompat - PackageValidation] Enable baseline assembly references
If assembly references aren't provided for the baseline package via the
MSBuild task or the CLI frontend, use the current package's assembly
references as an approximation.

As current assembly references are grouped by TFMs, use the nearest
compatible TFM for which assembly references are provided. This
guarantees that assembly references from current are used even if TFMs
don't match.