Fixes https://github.com/dotnet/sdk/issues/28746
NuGet APIs return TargetFramework information per ContentItem parsed
from a package. PackageValidation compared that TFM information as a
short folder name string with the TFM that the project targets when
PackageValidation is invoked from the msbuild frontend.

That TFM isn't indentical with what NuGet returns as the
TargetPlatformMoniker could be different, i.e. a default version is
appended. Also, that logic doesn't when TFMs are used aliases.

Instead compare the TargetFrameworkMoniker and the TargetPlatformMoniker
and pass them in to the msbuild frontend and require it in the CLI
frontend.