Allow PublishSingleFile and PublishAot specified simultaneously
We made native AOT targets error out if these are specified at the same time due to bad failure mode. But running through this, it appears to be mostly harmless except for this task.

So proposing taking this and rolling back https://github.com/dotnet/runtime/pull/73995. But no hard feelings if people feel otherwise.

We might want to eventually set `PublishSingleFile` when `PublishAot` is set, same way we do for `PublishTrimmed`. Some of the SDK behaviors look relevant. Others don't. I don't have a good sense on this one.