When transforming web.config check if target framework is .NET Framework.
This will avoid the aspNetCore elements from being added to the web.config when the user targets .NET Framework in an SDK style project and as such avoid a possible runtime error when IIS tries to use the AspNetCore module from the web.config inside of a .NET Framework targeted application and fail. The risk from this should be minimal because of the fact that non-SDK style projects do the same exact thing and they run as-is.

Fixes: https://github.com/dotnet/sdk/issues/34595.