Skip getting Windows GUI bit on non-Windows in AppHostShellShimMaker (#35600)
`AppHostShellShimMaker` had an implicit assumption that `ResourceUpdater.IsSupportedOS()` was tied to Windows. That is no longer the case as of RC1, resulting in `PEUtils.GetWindowsGraphicalUserInterfaceBit` being unnecessarily called on non-Windows when creating a shim.

This doesn't address the underlying issue around the function not properly handling endianness, but it brings back the previous behaviour where the problematic function isn't called when not needed.