Add ProjectCapability for Native AOT properties. (#35710)
Fixes [AB#1831643](https://devdiv.visualstudio.com/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_workitems/edit/1831643)

Description:
In .NET 8 we introduced options to customize the publish and build process to configure Native AOT properties into the Project Properties UI (see image below). To control which projects should surface these, the "NativeAOTProperties" project capability is used in the project system to control their visibility.

The idea is that we want to set this capability to true for projects targeting .NET 8 or higher in the .NET SDK side. The exceptions at the moment would be class libraries, WPF and WinForms projects, which are already filtered in the PR using the property condition. Other project types, like MAUI or Web projects, will be able to disable this capability by setting the "ShowNativeAOTProperties" to false in their props file. The change in this PR is needed so that other teams can override the default value in their .props file.

Risk:
Low, this change does not affect other than the visibility of the set of properties for Native AOT.
The workaround for this is to manually set them in the project file, if users want to edit these properties.

Validation:
Manual

Regression:
Not a regression