SDK resolver perf: Don't check current process for dotnet.exe on .NET Framework (#35890)
When evaluating projects in Visual Studio, more than half of the CPU spent in SDK resolution is in EnvironmentProvider.GetCurrentProcessPath(). Making this call in the .NET Framework version of the resolver does not make sense as there is no Framework dotnet.exe. This PRs reorders the checks so that EnvironmentProvider.GetCurrentProcessPath() does not run except for failure cases, which speeds up evaluation of the OrchardCore solution by ~18%.

The Framework Process.MainModule getter is slow. In this trace the total CPU spent in evaluation was 9459 ms.