Fix: Validate attribute type defines when excluded
Fixes half of https://github.com/dotnet/sdk/issues/33388

When attributes are configured to be excluded from APICompat
compatibility validation, the attribute definitions themselves were also
incorrectly excluded, not just their usage on members.

This fixes that by storing the attribute exclusion filter in a separate
field in the rule settings.

Added a test that excercises that behavior.