Change the default for 'EnforceCodeStyleInBuild' to true with AnalysisLevel 9.0 or greater
Back when we introduced CodeStyle NuGet package and build-time enforcement of IDE CodeStyle analyzers, it was an experimental feature, and hence the MSBuild property `EnforceCodeStyleInBuild` guarding this feature was defaulted to `false`. Now that it has been few release cycles with this package and feature, we will default this property to `true`. The new default value will only apply for AnalysisLevel/TFM 9.0 or greater to avoid any breaking changes in existing projects that configure IDExxxx to warning/error severity in their editorconfig, but do not set this property and hence see warnings/errors in the IDE live analysis, but not on build.

NOTE: This change has been approved in the offline email discussion with Roslyn IDE and compiler teams.