Add SkipGetTargetFrameworkProperties to redist.csproj
One of the commits for [this PR](https://github.com/dotnet/installer/pull/17942) failed.  On investigation, the dotnet-toolset-internal zip was missing a lot of Microsoft.NET.Build.Extensions files.  Looking at the binlog for the dotnet/sdk build that produced this zip revealed that the `CopyAdditionalFilesToLayout` target from Microsoft.NET.Build.Extensions.Tasks.csproj ran after the `PublishMSBuildExtensions` target from redist.csproj.  So the target which was laying out some of these files was running after the target that consumed them in another project.

This appears to be because the `CopyAdditionalFilesToLayout` target only runs in the project's outer build.  However, project references by default only build the inner build for the best matching target framework of the referenced project, so even though there was a project reference from redist.csproj to Microsoft.NET.Build.Extensions.Tasks.csproj, it wasn't ensuring the necessary target ordering.  Adding `SkipGetTargetFrameworkProperties="true"` to the project reference should cause the outer build of the referenced project to run before the referencing project is built.