Base version (#37625)
The version of SaveInstallStateManifestVersions in src/Cli/dotnet/commands/dotnet-workload/install/NetSdkMsiInstallerClient.cs inadvertently just called itself with the same arguments, leading to a StackOverflowException every time it's called. This resolves that issue by making the base version public and just going straight to that.

It also includes some cleanup for creating InstallStateContents.